requirements.txt
alembic==1.4.2
astroid==2.4.1
autopep8==1.5.3
cachelib==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
greenlet==0.4.15
gunicorn==18.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
monotonic==1.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
requests==2.25.0
oauthlib==3.1.0
pyOpenSSL==20.0.1
